segment: Make decoration of the preedit buffer visually distinguishable
The preedit buffer that is displayed to the user includes decoration
such as the cursor to mark the input location and brackets to mark
the current segment. However, the characters used for decoration may
also be part of the user input, making it impossible to distinguish
between the user input and the decoration.

This commit modifies the `segment_get_input_decorated()` method so
that it renders the decoration and user input in different colors,
allowing the two to be distinguished.