japanese: Probe other conjugations if one did not match
Because of the way return values are propagated and checked by
`japanese_unconjugate()` and its callees, the function does not
try to probe other conjugations if one did not match.

This commit modifies `japanese_unconjugate()` and its callees so
that `-ENOENT` is used to represent the case that a conjugation
did not match, and that all conjugations will be probed, unless
an error occurred.