trie: Add reverse insertion and lookup methods
Tries can be used to efficiently identify conjugation candidates for a
Japanese input when the key is reversed. However, this puts the burden
of reversing the keys on the caller, making the trie awkward to use in
this situation.

This commit adds the two methods `trie_get_values_reverse()` and
`trie_insert_reverse()` that lookup data from a trie, and insert data
into a trie, respectively, using the key in reverse.