trie: Make `_trie_append_to_array()` return the new array size
Some callers of `_trie_append_to_array()` need to know the size of
the array after appending succeeded, but the method returns zero
upon success. Thus, a separate call to determine the size is needed.

This commit modifies the `_trie_append_to_array()` method so that
it returns the new size of the array instead of zero upon success.