trie: Don't count elements twice in `trie_get_values_reverse()`
When the `trie_get_values_reverse()` method returns elements from
the current node and subnodes, it adds the count for the current
node to the return value of the recursive call, even though the
latter includes the count of the former.

This commit modifies the `trie_get_values_reverse()` method so that
it does not count elements twice.