segment: Implement browsing of candidates
The `segment_get_input_decorated()` method concatenates all input
candidates into a string, which is then displayed. Because the
string is usually longer than the area where it is displayed, this
prevents more than a certain number of candidates from being
displayed.

This commit modifies the `segment_get_input_decorated()` method so
that it displays only the 10 candidates that are around the current
selection, allowing the user to browse through all candidates.